requirements.txt
streamlit
boto3
botocore
requests
python-dotenv 
fastapi
uvicorn
openai>=1.0.0 